Removed cast to double as this is not needed for cluster size in col/row (in contrast to total cluster size)

+++ src/modules/AnalysisDUT/AnalysisDUT.cpp
@@ -375,8 +375,8 @@ StatusCode AnalysisDUT::run(std::shared_ptr<Clipboard> clipboard) {
 
                 clusterSizeAssoc->Fill(static_cast<double>(cluster->size()));
                 clusterSizeAssocNorm->Fill(static_cast<double>(cluster->size()));
-                clusterSizeRowAssoc->Fill(static_cast<double>(cluster->rowWidth());
-                clusterSizeColAssoc->Fill(static_cast<double>(cluster->columnWidth());
+                clusterSizeRowAssoc->Fill(cluster->rowWidth());
+                clusterSizeColAssoc->Fill(cluster->columnWidth());
 
                 // Fill in-pixel plots: (all as function of track position within pixel cell)
                 if(is_within_roi) {
